Add jest test for flowchart script
