+++ package.json
@@ -8,7 +8,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
+    "test:watch": "jest --watch",
     "prepare": "husky install"
   },
   "repository": {
@@ -35,13 +36,16 @@
     "csv-parse": "^5.0.4",
     "formik": "^2.2.9",
     "jspdf": "^2.4.0",
-    "next": "11.1.2",
+    "next": "^12.1.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "styled-components": "^5.3.3",
     "yup": "0.32.11"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^12.1.4",
+    "@types/jest": "^27.4.1",
     "@types/react": "17.0.24",
     "@types/styled-components": "^5.1.17",
     "@typescript-eslint/eslint-plugin": "^5.3.1",
@@ -51,6 +55,7 @@
     "eslint-config-next": "11.1.2",
     "eslint-config-prettier": "^8.3.0",
     "husky": "^7.0.4",
+    "jest": "^27.5.1",
     "lint-staged": "^11.2.6",
     "prettier": "2.4.1",
     "typescript": "4.4.3"
